refactor: extract cluster fixtures from tests/conftest.py
Move cluster-domain fixtures out of the monolithic tests/conftest.py
into domain-specific modules under tests/fixtures/cluster/:
auth, binaries, cpu, infrastructure, namespaces, nodes, sanity, utilities.
New modules are registered in the root pytest_plugins list.

Also moves scaled_deployment (observability-specific) to
tests/observability/conftest.py.

Incidental fixes required by the move:
- skip_if_no_common_modern_cpu: added nodes_cpu_architecture as an
  explicit fixture parameter (original referenced the fixture function
  object as a free variable — always truthy, would become a NameError
  after the move).
- namespace fixture: replaced string-split path logic with
  pathlib.Path.resolve() + Path.relative_to() to correctly derive the
  test-relative path after the file moved from tests/ to
  tests/fixtures/cluster/, and to handle symlinked/non-canonical
  checkouts.

Part of CNV-80951.
