+++ tests/observability/conftest.py
@@ -4,6 +4,7 @@ import pytest
 from ocp_resources.ssp import SSP
 
 from utilities.hco import ResourceEditorValidateHCOReconcile
+from utilities.infra import scale_deployment_replicas
 
 LOGGER = logging.getLogger(__name__)
 
@@ -17,5 +18,16 @@ def paused_ssp_operator(admin_client, hco_namespace, ssp_resource_scope_class):
         admin_client=admin_client,
         patches={ssp_resource_scope_class: {"metadata": {"annotations": {"kubevirt.io/operator.paused": "true"}}}},
         list_resource_reconcile=[SSP],
+    ):
+        yield
+
+
+@pytest.fixture()
+def scaled_deployment(request, hco_namespace, admin_client):
+    with scale_deployment_replicas(
+        deployment_name=request.param["deployment_name"],
+        replica_count=request.param["replicas"],
+        namespace=hco_namespace.name,
+        client=admin_client,
     ):
         yield
